refactor: 요청 처리 구조 개선
- 무엇을 처리할지에 관한 로직을 카타리나로 이전하여 Processor는 HTTP 요청 메시지를 파싱하여 전달하고 응답을 반환하는 책임만 가지도록 구현
- 요청 매핑 객체 추가
